feat(cli): adds help section; breaks main in diferent files
