+++ cmd/fakeapi/main.go
@@ -1,51 +1,24 @@
 package main
 
 import (
-	"errors"
-	"flag"
 	"fmt"
-	stdlog "log"
-	"os"
-
-	log "github.com/sirupsen/logrus"
+	"log"
 
 	"fakeapi/pkg/route"
 	"fakeapi/pkg/server"
 )
 
-type cliArgs struct {
-	FilePath string
-	Port     int16
-}
-
-func newCLIArgs() (*cliArgs, error) {
-	port := flag.Int("port", 8080, "Port to run the server on")
-	flag.Parse()
-
-	params := flag.Args()
-	if len(params) < 1 {
-		return nil, errors.New("Inform the needed parameters")
-	}
-
-	dir, err := os.Getwd()
-	if err != nil {
-		return nil, err
-	}
-	filePath := fmt.Sprintf("%s/%s", dir, params[0])
-
-	return &cliArgs{FilePath: filePath, Port: int16(*port)}, nil
-}
-
 func main() {
 	setupLog()
 	cliArgs, err := newCLIArgs()
 	if err != nil {
-		stdlog.Fatal(err)
+		handleCLIErrors(err)
+		return
 	}
 
 	endpoints, err := route.NewEndpointsFromFile(cliArgs.FilePath)
 	if err != nil {
-		stdlog.Fatal(err)
+		log.Fatal(err)
 	}
 
 	server := server.NewServer(
@@ -55,12 +28,21 @@ func main() {
 
 	err = server.Start()
 	if err != nil {
-		stdlog.Fatal(err)
+		log.Fatal(err)
 	}
 }
 
-func setupLog() {
-	log.SetFormatter(&log.TextFormatter{})
-	log.SetOutput(os.Stdout)
-	log.SetLevel(log.InfoLevel)
+func handleCLIErrors(err error) {
+	if err == nil {
+		return
+	}
+
+	fmt.Println(err.Error())
+
+	if err == ErrNoParams {
+		printHelp()
+		return
+	}
+
+	log.Fatal(err)
 }
